I have a working video player.
The next challenge is getting the database to work.

I realized I could add a way to jump to a specific part of the video, but for now, the important thing is that the video plays.

I had to deal with some stuff with videoview1 click event being added, but its sorted out now. Basically, when I clicked on videoview 1 Visual Studio created a click event and then when I ran the program it had an error because that method was not implemented, but I didn't want that, so I removed that click event and it was fine and ran normally.